Temporarily remove tomthorogood.uk test cases
I've been having problems with Cloudflare and Google Cloud (Firebase)
so I've had to temporarily disable Cloudflare's proxy which seems to
have caused this domain to be removed from the preload list. Temporarily
disabling these test cases allows the tests to pass and the list to be
updated automatically once more.